% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_anova.test}
\alias{gen_anova.test}
\title{Generate sample data for one-way ANOVA simulations}
\usage{
gen_anova.test(n, k, f, n.ratios = rep(1, k), means = NULL, sds = NULL, ...)
}
\arguments{
\item{n}{sample size per group}

\item{k}{number of groups}

\item{f}{Cohen's f effect size}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{means}{(optional) vector of means. When specified the input \code{f} is ignored}

\item{sds}{(optional) vector of SDs. When specified the input \code{f} is ignored}

\item{...}{additional arguments (not used)}
}
\value{
returns a data.frame with the variables 'DV' and 'group'
}
\description{
Generates continuous multi-sample data to be analysed by
a one-way ANOVA. Samples are generated given conditional
observations are normally distributed and have
have equal variance by default, however these may be modified.
}
\examples{

# n=50 in 3 groups, "medium" effect size
gen_anova.test(50, k=3, f=.25)

# explicit means/sds
gen_anova.test(50, 3, means=c(0,0,1), sds=c(1,2,1))

}
\seealso{
\code{\link{p_anova.test}}
}
