% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_chisq.test}
\alias{gen_chisq.test}
\title{Generate multinomial data for chi-squared test}
\usage{
gen_chisq.test(n, P, ...)
}
\arguments{
\item{n}{sample size per group}

\item{P}{specific power configuration, specified as a numeric vector or matrix}

\item{...}{additional arguments (not used)}
}
\value{
a numeric vector or matrix, depending on the supplied \code{P} class
}
\description{
Generates multinomial data suitable for analysis with
\code{\link{chisq.test}}.
}
\examples{

# vector of explicit probabilities (goodness of fit test)
gen_chisq.test(100, P0 = c(.25, .25, .25, .25),
                     P = c(.6, .2, .1, .1))

# matrix of explicit probabilities (two-dimensional test of independence)
gen_chisq.test(100, P0 = matrix(c(.25, .25, .25, .25), 2, 2),
                     P = matrix(c(.6, .2, .1, .1),2,2))

}
\seealso{
\code{\link{p_chisq.test}}
}
