% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_lm}
\alias{gen_lm}
\title{Generate data samples for global linear regression model simulation}
\usage{
gen_lm(n, k, R2, R2_0 = 0, k.R2_0 = 0, R2.resid = 1 - R2, fixed.X = TRUE, ...)
}
\arguments{
\item{n}{sample size}

\item{k}{number of IVs}

\item{R2}{R-squared effect size}

\item{R2_0}{null hypothesis for R-squared}

\item{k.R2_0}{number of IVs associated with the null hypothesis model}

\item{R2.resid}{residual R-squared value, typically used when comparing
nested models when fit sequentially (e.g., comparing model A vs B when
model involves the structure A -> B -> C)}

\item{fixed.X}{logical; should the IVs be considered fixed or random?}

\item{...}{additional arguments (not used)}
}
\value{
a data.frame with the dependent variable in the left-most column
with the name 'y', while all other columns reflect the independent variables
}
\description{
Data generation for linear regression model using fixed or random
independent variables.
}
\examples{

# 5 fixed IVs, R^2 = .1, sample size of 95
gen_lm(n=95, R2=.1, k=5)

}
\seealso{
\code{\link{p_lm}}
}
