% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_mcnemar.test}
\alias{gen_mcnemar.test}
\title{Generate multinomial data for McNemar test}
\usage{
gen_mcnemar.test(n, prop, ...)
}
\arguments{
\item{n}{total sample size}

\item{prop}{two-dimensional matrix of proportions/probabilities}

\item{...}{additional arguments (not used)}
}
\value{
a matrix of counts
}
\description{
Generates two-dimensional sample data for McNemar test
from a multinomial distribution.
}
\examples{

# from ?mcnemar.test
Performance <- matrix(c(794, 86, 150, 570),
		   nrow = 2,
		   dimnames = list("1st Survey" = c("Approve", "Disapprove"),
		               "2nd Survey" = c("Approve", "Disapprove")))
(prop <- prop.table(Performance))

# table of counts given prop input
gen_mcnemar.test(n=sum(Performance), prop=prop)


}
\seealso{
\code{\link{p_mcnemar.test}}
}
