% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_mediation}
\alias{gen_mediation}
\title{Generates three-variable mediation analysis data}
\usage{
gen_mediation(n, a, b, cprime, dichotomous.X = FALSE, ...)
}
\arguments{
\item{n}{total sample size unless \code{dichotomous.X = TRUE}, in which the
value represents the size per group}

\item{a}{regression coefficient for the path X -> M}

\item{b}{regression coefficient for the path M -> Y}

\item{cprime}{partial regression coefficient for the path X -> Y}

\item{dichotomous.X}{logical; should the X variable be generated as though it
were dichotomous? If TRUE then \code{n} represents the sample size per group}
}
\value{
a data.frame with the variables X, M, and Y for the independent variable,
mediator, and dependent variable, respectively
}
\description{
Generates continuous X and Y data and continuous or dichotomous X data
in a three variable mediation model, assuming Gaussian errors.
}
\examples{

gen_mediation(50, a=sqrt(.35), b=sqrt(.35), cprime=.39)
gen_mediation(50, a=sqrt(.35), b=sqrt(.35), cprime=.39, dichotomous.X=TRUE)

}
\seealso{
\code{\link{p_mediate}}
}
