% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_mvtnorm}
\alias{gen_mvtnorm}
\title{Generate multivariate normal data}
\usage{
gen_mvtnorm(n, mean = numeric(nrow(sigma)), sigma, ...)
}
\arguments{
\item{n}{sample size}

\item{mean}{vector of means. Defaults to a vector of 0's}

\item{sigma}{covariance matrix, though if all diagonal values equal 1
then understood as a correlation matrix}

\item{...}{additional arguments (not used)}
}
\value{
a \code{matrix} with two-columns and \code{n} rows
}
\description{
Generates correlated X-Y data from a bivariate normal distribution.
}
\examples{

sigma <- matrix(c(1,.2,.3,.2,1,.5,.3,.5,1), 3, 3)
dat <- gen_mvtnorm(1000, sigma=sigma)
pairs(dat)

}
\seealso{
\code{\link{p_r}} and \code{\link{p_r.cat}}
}
