% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{gen_r}
\alias{gen_r}
\title{Generate bivariate normal data}
\usage{
gen_r(n, r, ...)
}
\arguments{
\item{n}{sample size}

\item{r}{correlation}

\item{...}{additional arguments (not used)}
}
\value{
a \code{matrix} with two-columns and \code{n} rows
}
\description{
Generates correlated X-Y data from a bivariate normal distribution.
}
\examples{

dat <- gen_r(1000, r=.9)
plot(dat)

}
\seealso{
\code{\link{p_r}}
}
