% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{gen_t.test}
\alias{gen_t.test}
\title{Generate data for paired, one-sample, and two-sample t.test procedures}
\usage{
gen_t.test(
  n,
  d,
  n2_n1 = 1,
  r = NULL,
  type = c("two.sample", "one.sample", "paired"),
  raw_info = list(means = NA, sds = NA),
  ...
)
}
\arguments{
\item{n}{sample size per group, assumed equal across groups}

\item{d}{Cohen's standardized effect size \code{d}}

\item{n2_n1}{allocation ratio reflecting the same size ratio.
Default of 1 sets the groups to be the same size. Only applicable
when \code{type = 'two.sample'}}

\item{r}{(optional) instead of specifying \code{d} specify
a point-biserial correlation. Internally this is transformed
into a suitable \code{d} value for the power computations}

\item{type}{type of data structure to generate; can be \code{'two.sample'},
\code{'one.sample'}, or \code{'paired'}}

\item{raw_info}{(optional) list of mean and SD inputs for each group,
each specified as a vector. When specified the input \code{d} is ignored}

\item{...}{additional arguments (not used)}
}
\value{
a \code{data.frame} with the columns \code{'DV'} and \code{'group'}
}
\description{
Generates one or two sets of continuous data group-level data
according to Cohen's effect size 'd'. The data are generated such that
the conditional observations are normally distributed and have
have equal variance, however some of these properties may be modified.
}
\examples{

# two-sample data with 100 observations each
df <- gen_t.test(n=100, d=.5)
head(df)
with(df, tapply(DV, group, mean))

}
\seealso{
\code{\link{p_t.test}}
}
