% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_var.test}
\alias{gen_var.test}
\title{Generate sample data for variance test}
\usage{
gen_var.test(n, sds, n.ratios = rep(1, length(sds)), ...)
}
\arguments{
\item{n}{sample size per group, assumed equal across groups}

\item{sds}{a vector of standard deviations to use for each group}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{...}{additional arguments (not used)}
}
\value{
a data.frame with the variables 'DV' and (potentially) 'group' when
constructing multi-sample data
}
\description{
Generates one or or more sets of continuous data group-level data
to perform a variance test.
}
\examples{

# one sample
gen_var.test(100, sds=10)

# three sample
gen_var.test(100, sds=c(10, 9, 11))

}
\seealso{
\code{\link{p_var.test}}
}
