% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getLastSpower}
\alias{getLastSpower}
\title{Get previously evaluated Spower execution}
\usage{
getLastSpower()
}
\value{
the last object returned from \code{\link{Spower}} or \code{\link{SpowerBatch}}
}
\description{
If the result of \code{\link{Spower}} or \code{\link{SpowerBatch}} was not stored into
an object this function will retrieve the last evaluation.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
