% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{p_2r}
\alias{p_2r}
\title{Compare two correlations}
\usage{
p_2r(
  n,
  r.xy1,
  r.xy2,
  r.xz1,
  r.xz2,
  r.yz1,
  r.yz2,
  r.xw1,
  r.xw2,
  r.yw1,
  r.yw2,
  r.zw1,
  r.zw2,
  n2_n1 = 1,
  two.tailed = TRUE,
  type = c("independent", "overlap", "nonoverlap"),
  test = "fisher1925"
)
}
\description{
Compare two correlations
}
\examples{

# independent (same x-y pairing across groups)
p_2r(100, r.xy1=.5, r.xy2=.6)

# overlap (same y, different xs)
p_2r(100, r.xy1=.5, r.xy2=.7,
          r.xz1=.3, r.xz2=.3,
          r.yz1=.2, r.yz2=.2, type = 'overlap')

# nonoverlap (different ys, different xs)
p_2r(100, r.xy1=.5, r.xy2=.6,
          r.xz1=.3, r.xz2=.3,
          r.yz1=.2, r.yz2=.2,
          r.xw1=.2, r.xw2=.2,
          r.yw1=.4, r.yw2=.4,
          r.zw1=.2, r.zw2=.2,
          type = 'nonoverlap')


}
