% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ansari.test.R
\name{p_ansari.test}
\alias{p_ansari.test}
\title{p-value from Ansari-Bradley Test simulation}
\usage{
p_ansari.test(
  n,
  scale,
  n2_n1 = 1,
  two.tailed = TRUE,
  exact = NULL,
  parent = function(n, ...) rnorm(n),
  ...
)
}
\arguments{
\item{n}{sample size per group}

\item{scale}{the scale to multiply the second group by (1 reflects equal scaling)}

\item{n2_n1}{sample size ratio}

\item{two.tailed}{logical; use two-tailed test?}

\item{exact}{a logical indicating whether an exact p-value should be computed}

\item{parent}{data generation function (default assumes Gaussian shape). Must be
population mean centered}

\item{...}{additional arguments to pass to simulation functions (if used)}
}
\description{
Simulates data given one or two parent distributions and
returns a p-value testing that the scale of the type distributions are the same.
Default implementation uses Gaussian distributions, however the
distribution function may be modified to
reflect other populations of interest. Uses \code{\link{ansari.test}} for the analysis.
}
\examples{

# n=30 per group,
#  Distributions Gaussian with sd=1 for first group and sd=2 for second
p_ansari.test(30, scale=2)

# compare chi-squared distributions
parent <- function(n, df, ...) rchisq(n, df=df) - df
p_ansari.test(30, scale=2, parent=parent, df=3)

\dontrun{
  # empirical power of the experiments
  p_ansari.test(30, scale=2) |> Spower()
  p_ansari.test(30, scale=2, parent=parent, df=3) |> Spower()

}
}
