% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_chisq.test}
\alias{p_chisq.test}
\title{Proportion test simulation and p-value}
\usage{
p_chisq.test(n, w, df, raw_info = list(P0 = NA, P = NA))
}
\arguments{
\item{n}{sample size per group}

\item{prop}{sample probability/proportions of success.
If a vector with two-values or more elements are supplied then
a multi-samples test will be used}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{pi}{probability of success to test against (default is .5). Ignored
for two-sample tests}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}
}
\value{
a single p-value
}
\description{
Generates single and multi-sample data
for proportion tests and return a p-value. Uses \code{\link{binom.test}}
for one-sample applications and \code{\link{prop.test}} otherwise.
}
\examples{

p_chisq.test(100, raw_info = list(P0 = c(.25, .25, .25, .25),
                                  P = c(.6, .2, .1, .1)))

# works, but logic seems odd
p_chisq.test(100, raw_info = list(P0 = matrix(c(.25, .25, .25, .25), 2, 2),
                                  P = matrix(c(.6, .2, .1, .1),2,2)))

if(FALSE){
    # compare simulated results to pwr package


}

}
