% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{p_lm}
\alias{p_lm}
\title{Linear regression model and p-value}
\usage{
p_lm(n, R2, k, R2_0 = 0, k.R2_0 = 0, R2.resid = 1 - R2, fixed.X = TRUE)
}
\arguments{
\item{n}{sample size}

\item{R2}{R-squared effect size}

\item{k}{number of IVs}

\item{R2_0}{null hypothesis for R-squared}

\item{k.R2_0}{number of IVs associated with the null hypothesis model}

\item{R2.resid}{residual R-squared value, typically used when comparing
nested models when fit sequentially (e.g., comparing model A vs B when
model involves the structure A -> B -> C)}

\item{fixed.X}{logical; should the IVs be considered fixed or random?}
}
\description{
p-values associated with linear regression model using fixed or random
independent variables.
}
\examples{

# 5 fixed IVs, R^2 = .1, sample size of 95
p_lm(n=95, R2=.1, k=5)

}
