% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_extras.R
\name{p_mediation}
\alias{p_mediation}
\title{Three-variable mediation analysis and p-value}
\usage{
p_mediation(
  n,
  a,
  b,
  cprime,
  dichotomous.X = FALSE,
  two.tailed = TRUE,
  method = "wald"
)
}
\arguments{
\item{n}{total sample size}

\item{a}{regression coefficient for the path X -> M}

\item{b}{regression coefficient for the path M -> Y}

\item{cprime}{partial regression coefficient for the path X -> Y}

\item{dichotomous.X}{logical; should the X variable be generated as though it
were dichotomous? If TRUE then \code{n} represents the sample size per group}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{method}{type of inferential method to use. Default uses the Wald
(a.k.a., Sobel) test}
}
\value{
a single p-value
}
\description{
Simple 3-variable mediation analysis simulation to test the hypothesis that
X -> Y is mediated by the relationship X -> M -> Y. Currently,
M and Y are assumed to be continuous variables with Gaussian errors, while
X may be continuous or dichotomous.
}
\examples{

# joint test H0: a*b = 0
p_mediation(50, a=sqrt(.35), b=sqrt(.35), cprime=.39)
p_mediation(50, a=sqrt(.35), b=sqrt(.35), cprime=.39, dichotomous.X=TRUE)

if(FALSE){

  # power to detect mediation
  Spower(p_mediation, n=50, a=sqrt(.35), b=sqrt(.35),
         cprime=.39, parallel=TRUE, replications=1000)

  # sample size for .95 power
  Spower(p_mediation, n=NA, a=sqrt(.35), b=sqrt(.35),
         cprime=.39, power=.95, interval=c(50, 300),
         parallel=TRUE)


}

}
