% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_prop.test}
\alias{p_prop.test}
\title{Proportion test simulation and p-value}
\usage{
p_prop.test(
  n,
  prob,
  pi = 0.5,
  n.ratios = rep(1, length(prob)),
  two.tailed = TRUE
)
}
\arguments{
\item{n}{sample size per group}

\item{prob}{sample probability of success. If a vector with two-values
is supplied then a two-samples test will be used}

\item{pi}{probability of success to test against (default is .5). Ignored
for two-sample tests}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}
}
\value{
a single p-value
}
\description{
Proportion test simulation and p-value
}
\examples{

# 50 observations, test against pi = .5
p_prop.test(50, prob=.65)

# two sample test
p_prop.test(50, prob=c(.5, .65))

# two sample test, unequal ns
p_prop.test(50, prob=c(.5, .65))

# three sample test, group2 twice as large as others
p_prop.test(50, prob=c(.5, .65, .7), n.ratios=c(1,2,1))

}
