% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{p_r.cat}
\alias{p_r.cat}
\title{Polychoric and polyserial simulation and p-value}
\usage{
p_r.cat(
  n,
  r,
  tauX,
  rho = 0,
  tauY = NULL,
  ML = TRUE,
  two.tailed = TRUE,
  score = TRUE
)
}
\arguments{
\item{n}{sample size}

\item{r}{correlation prior to the discretization (recovered via the
polyserial/polychoric estimates)}

\item{tauX}{intercept parameters used for discretizing the X variable}

\item{rho}{population coefficient to test against}

\item{tauY}{intercept parameters used for discretizing the Y variable. If
missing a polyserial correlation will be estimated, otherwise a
tetrachoric/polychoric correlation will be estimated}

\item{ML}{logical; use maximum-likelihood estimation?}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{score}{logical; should the SE be based at the null hypothesis (score test)
or the ML estimate (Wald test)?}
}
\value{
a single p-value
}
\description{
Generates correlated X-Y data and returns a p-value to assess the null
of no correlation in the population. The X-Y data are generated
assuming a multivariate normal distribution and subsequently discretized
for one or both of the variables.
}
\examples{

# 100 observations, .5 correlation, tetrachoric estimate
p_r.cat(100, r=.5, tauX=0, tauY=1)

# polyserial estimate (Y continuous)
p_r.cat(50, r=.5, tauX=0)

}
