% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_r}
\alias{p_r}
\title{Correlation simulation and p-value}
\usage{
p_r(n, r, rho = 0, method = "pearson", two.tailed = TRUE)
}
\arguments{
\item{n}{sample size}

\item{r}{correlation}

\item{rho}{population coefficient to test against using
\code{\link[car]{linearHypothesis}}}

\item{method}{method to use to compute the correlation
(see \code{\link{cor.test}}). Only used when \code{rho = 0}}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}
}
\value{
a single p-value
}
\description{
Function generates correlated X-Y data and returns a p-value to assess the null
of no correlation in the population. The X-Y data are generated
assuming a multivariate normal distribution.
}
\examples{

# 50 observations, .5 correlation
p_r(50, r=.5)
p_r(50, r=.5, method = 'spearman')

# test against constant other than rho = .6
p_r(50, .5, rho=.60)

}
