% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCurve.R
\name{powerCurve}
\alias{powerCurve}
\title{Draw power curve from simulation functions}
\usage{
powerCurve(
  sim,
  varying,
  ...,
  interval = NULL,
  power = NA,
  sig.level = 0.05,
  replications = 10000,
  integer,
  prior = NULL,
  parallel = FALSE,
  cl = NULL,
  ncores = parallelly::availableCores(omit = 1L),
  predCI = 0.95,
  predCI.tol = 0.01,
  verbose = TRUE,
  check.interval = FALSE,
  maxiter = 150,
  wait.time = NULL,
  control = list()
)
}
\arguments{
\item{sim}{function that both creates the data and returns a single
p-value for the analysis of interest}

\item{varying}{vector of values to substitute into the missing \code{...}
terms. Requires \code{power} to be set to \code{NA}}

\item{...}{a set of conditions to use in the simulation that must match the
arguments in the function \code{sim}. See \code{\link{Spower}}}

\item{interval}{search interval to use when \code{\link{SimSolve}} is required.
Can be a vector of length two to apply the same interval across
the \code{varying} information or a \code{matrix} with two columns
to apply intervals on a per-row basis}

\item{power}{power level to use. If set to \code{NA} then the empirical power
will be estimated given the fixed \code{...} input; otherwise,
can be specified as a vector to solve the missing elements in
\code{...}}

\item{sig.level}{see \code{\link{Spower}}}

\item{replications}{see \code{\link{Spower}}}

\item{integer}{see \code{\link{Spower}}}

\item{prior}{see \code{\link{Spower}}}

\item{parallel}{see \code{\link{Spower}}}

\item{cl}{see \code{\link{Spower}}}

\item{ncores}{see \code{\link{Spower}}}

\item{predCI}{see \code{\link{Spower}}}

\item{predCI.tol}{see \code{\link{Spower}}}

\item{verbose}{see \code{\link{Spower}}}

\item{check.interval}{see \code{\link{Spower}}, though is set to \code{FALSE}
by default instead}

\item{maxiter}{see \code{\link{Spower}}}

\item{wait.time}{see \code{\link{Spower}}}

\item{control}{see \code{\link{Spower}}}
}
\value{
a ggplot2 object
}
\description{
Draws power curves that either a) estimate the power given a
set of varying conditions or b) solves a set of root conditions
given fixed values of power. Confidence/prediction intervals are
included in the output to reflect the estimate uncertainties.
}
\examples{
\dontrun{

# estimate power given varying sample sizes
gg <- powerCurve(p_t.test, varying=c(30, 90, 270, 550), n=NA, d=0.2,
 		   replications=1000)

# Because output is a ggplot2 object can be modified
gg + geom_text(aes(label=power), size=5, colour='red', nudge_y=.05) +
  ylab(expression(1-beta)) + theme_grey()

# using default precision (10000 replications). Parallel computations
# generally recommended in this case to save time
powerCurve(p_t.test, varying=c(30, 90, 270, 550), n=NA, d=0.2)

# alternatively, get information from last printed graphic instead of saving
gg <- last_plot()
gg + coord_flip() # flip coordinates

# estimate sample sizes given varying power
powerCurve(p_t.test, n=NA, d=0.2, interval=c(10, 1000),
		   power=c(.1, .25, .5, .75, .9), maxiter=30)

# estimate power varying d
powerCurve(p_t.test, varying=seq(.1, 1, by=.2), n=50, d=NA,
		   replications=1000)

# estimate d varying power
powerCurve(p_t.test, n=50, d=NA,
		   maxiter=30, interval=c(.01, 1),
		   power=c(.1, .25, .5, .75, .9))

}

}
\seealso{
\code{\link{Spower}}
}
