% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpowerCluster.R
\name{SpowerCluster}
\alias{SpowerCluster}
\title{Define a parallel cluster object to be used in internal functions}
\usage{
SpowerCluster(spec, remove = FALSE, ...)
}
\arguments{
\item{spec}{input that is passed to \code{parallel::makeCluster()}. If no input is given the
maximum number of available local cores minus 1 will be used.
Setting this to NULL will skip a new definition (allows \code{omp_threads} to be used independently)}

\item{remove}{logical; remove previously defined \code{SpowerCluster()}?}

\item{...}{additional arguments to pass to \code{parallel::makeCluster}}
}
\description{
This function defines a object that is placed in a relevant internal environment defined in Spower,
and stores the object internally so that it can be reused across multiple applications rather than
redefining a cluster object for each analysis. Note that this is performed automatically when
a call to \code{\link{Spower}} uses \code{parallel = TRUE}. To remove the internal object
for the purpose of redefining a cluster use the argument \code{remove = TRUE}.
}
\examples{

\dontrun{
if(interactive()){
  # use all available cores
  SpowerCluster()
  SpowerCluster(remove = TRUE)

}

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{parallel}
