% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{update_sig.level}
\alias{update_sig.level}
\title{Update power estimates given new sig.level without re-simulating}
\usage{
update_sig.level(x, sig.level, predCI = 0.95)
}
\arguments{
\item{x}{object returned from \code{\link{Spower}} where \code{power}
was estimated or the \code{bete_alpha} criteria were supplied}

\item{sig.level}{Type I error rate (alpha)}

\item{predCI}{confidence interval precision (see \code{\link{Spower}} for
similiar input)}
}
\description{
When a power was performed with \code{\link{Spower}} this
function can be used to
update the power estimate given an alternative Type I error rate
(alpha/sig.level) without re-simulating the experiment.
}
\examples{
\dontrun{

# estimate power using sig.level = .05 (default)
out <- Spower(p_t.test, n = 50, d = .5)

# update power estimate given sig.level=.01 and .20
update_sig.level(out, sig.level=.01)
update_sig.level(out, sig.level=.20)

}

}
