% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{updateCompromise}
\alias{updateCompromise}
\title{Update compromise analysis criteria without re-simulating}
\usage{
updateCompromise(x, beta_alpha)
}
\arguments{
\item{x}{object returned from \code{\link{Spower}} where \code{power}
was estimated or the \code{bete_alpha} criteria were supplied}

\item{beta_alpha}{Type II/Type I error ratio}
}
\description{
When a power or compromise analysis was performed in
\code{\link{Spower}} this function can be used to
update the compromise criteria without the need for re-simulating
the experiment.
}
\examples{
\dontrun{
# Solve beta/alpha ratio to specific error trade-off constant
out <- Spower(p_t.test, n = 50, d = .5, beta_alpha = 2)

# update beta_alpha criteria without re-simulating
updateCompromise(out, beta_alpha=4)

}

}
