% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim.R
\name{gen_prop.test}
\alias{gen_prop.test}
\title{Generate binomial data for proportion tests}
\usage{
gen_prop.test(n, prop, n.ratios = rep(1, length(prop)), ...)
}
\arguments{
\item{n}{sample size per group}

\item{prop}{sample probability/proportions of success.
If a vector with two-values or more elements are supplied then
a multi-samples test will be used. Matrices are also supported}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{...}{additional arguments (not used)}
}
\value{
an integer vector for one-sample data generation or
a 2xk matrix of counts for multi-sample problems
}
\description{
Generates single and multi-sample binomial data for proportion tests.
}
\examples{

# one sample, 50 observations
gen_prop.test(50, prop=.65)

# two-sample test
gen_prop.test(50, prop=c(.5, .65))

# two-sample test, unequal ns
gen_prop.test(50, prop=c(.5, .65), n.ratios = c(1,2))

# three-sample test, group2 twice as large as others
gen_prop.test(50, prop=c(.5, .65, .7), n.ratios=c(1,2,1))

# data for Fisher exact test
gen_prop.test(50, prop=matrix(c(.5, .65, .7, .5), 2, 2))


}
\seealso{
\code{\link{p_prop.test}}
}
