% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_chisq.test}
\alias{p_chisq.test}
\title{Chi-squared test simulation and p-value}
\usage{
p_chisq.test(n, w, df, correct = TRUE, raw_info = list(P0 = NA, P = NA))
}
\arguments{
\item{n}{sample size per group}

\item{w}{Cohen's w effect size}

\item{df}{degrees of freedom}

\item{correct}{logical; apply continuity correction?}

\item{raw_info}{list of raw information to generate multinomial data
under specific nulls and power configurations}
}
\value{
a single p-value
}
\description{
Generates multinomial data suitable for analysis with
\code{\link{chisq.test}}.
}
\examples{

# effect size w + df
p_chisq.test(100, w=.2, df=3)

# vector of explicit probabilities
p_chisq.test(100, raw_info = list(P0 = c(.25, .25, .25, .25),
                                  P = c(.6, .2, .1, .1)))

# matrix of explicit probabilities
p_chisq.test(100, raw_info = list(P0 = matrix(c(.25, .25, .25, .25), 2, 2),
                                  P = matrix(c(.6, .2, .1, .1),2,2)))

if(FALSE){
    # compare simulated results to pwr package

    P0 <- c(1/3, 1/3, 1/3)
    P <- c(.5, .25, .25)
    w <- pwr::ES.w1(P0, P)
    df <- 3-1
    pwr::pwr.chisq.test(w=w, df=df, N=100, sig.level=0.05)

    Spower(p_chisq.test, n=100, w=w, df=df)
    Spower(p_chisq.test, n=100, raw_info=list(P0=P0, P=P))

    # slightly differ (latter more conservative)
    pwr::pwr.chisq.test(w=w, df=df, power=.8, sig.level=0.05)
    Spower(p_chisq.test, n=NA, w=w, df=df,
           power=.80, interval=c(50, 200))

    # Spower more conservative even with large N
    pwr::pwr.chisq.test(w=.1, df=df, power=.95, sig.level=0.05)
    out <- Spower(p_chisq.test, n=NA, w=.1, df=df,
                 power=.95, interval=c(1000, 2000))
    summary(out)

}

}
