% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_lm.R
\name{p_lm}
\alias{p_lm}
\alias{gen_lm}
\title{p-value from global linear regression model simulation}
\usage{
p_lm(
  n,
  R2,
  k,
  R2_0 = 0,
  k.R2_0 = 0,
  R2.resid = 1 - R2,
  fixed.X = TRUE,
  gen_fun = gen_lm,
  ...
)

gen_lm(n, k, R2, R2_0 = 0, k.R2_0 = 0, R2.resid = 1 - R2, fixed.X = TRUE, ...)
}
\arguments{
\item{n}{sample size}

\item{R2}{R-squared effect size}

\item{k}{number of IVs}

\item{R2_0}{null hypothesis for R-squared}

\item{k.R2_0}{number of IVs associated with the null hypothesis model}

\item{R2.resid}{residual R-squared value, typically used when comparing
nested models when fit sequentially (e.g., comparing model A vs B when
model involves the structure A -> B -> C)}

\item{fixed.X}{logical; should the IVs be considered fixed or random?}

\item{gen_fun}{function used to generate the required discrete data.
Object returned must be a \code{data.frame}. Default uses \code{\link{gen_lm}}.
User defined version of this function must include the argument \code{...}}

\item{...}{additional arguments to be passed to \code{gen_fun}. Not used
unless a customized \code{gen_fun} is defined}
}
\description{
p-values associated with linear regression model using fixed or random
independent variables.
}
\examples{

# 5 fixed IVs, R^2 = .1, sample size of 95
p_lm(n=95, R2=.1, k=5)

}
\seealso{
\code{\link{gen_lm}}
}
