% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates_Gpower.R
\name{p_mcnemar.test}
\alias{p_mcnemar.test}
\title{McNemar test simulation and p-value}
\usage{
p_mcnemar.test(n, prop, two.tailed = TRUE, correct = TRUE)
}
\arguments{
\item{n}{total sample size}

\item{prop}{two-dimensional matrix of proportions/probabilities}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{correct}{logical; use continuity correction? Only applicable for
2x2 tables}
}
\value{
a single p-value
}
\description{
Generates two-dimensional sample data for McNemar test and
return a p-value. Uses \code{\link{mcnemar.test}}.
}
\examples{

# from ?mcnemar.test
Performance <- matrix(c(794, 86, 150, 570),
		   nrow = 2,
		   dimnames = list("1st Survey" = c("Approve", "Disapprove"),
		               "2nd Survey" = c("Approve", "Disapprove")))
(prop <- prop.table(Performance))

# one sample + test and resulting p-value
p_mcnemar.test(n=sum(Performance), prop=prop)

\dontrun{

# post-hoc power (not recommended)
Spower(p_mcnemar.test, n=sum(Performance), prop=prop)

}

}
