% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_prop.test}
\alias{p_prop.test}
\title{Proportion test simulation and p-value}
\usage{
p_prop.test(
  n,
  prop,
  pi = 0.5,
  n.ratios = rep(1, length(prop)),
  two.tailed = TRUE
)
}
\arguments{
\item{n}{sample size per group}

\item{prop}{sample probability/proportions of success.
If a vector with two-values or more elements are supplied then
a multi-samples test will be used}

\item{pi}{probability of success to test against (default is .5). Ignored
for two-sample tests}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}
}
\value{
a single p-value
}
\description{
Generates single and multi-sample data
for proportion tests and return a p-value. Uses \code{\link{binom.test}}
for one-sample applications and \code{\link{prop.test}} otherwise.
}
\examples{

# 50 observations, test against pi = .5
p_prop.test(50, prop=.65)

# two sample test
p_prop.test(50, prop=c(.5, .65))

# two sample test, unequal ns
p_prop.test(50, prop=c(.5, .65))

# three sample test, group2 twice as large as others
p_prop.test(50, prop=c(.5, .65, .7), n.ratios=c(1,2,1))

if(FALSE){
    # compare simulated results to pwr package

    h <- pwr::ES.h(0.5, 0.4)
    pwr::pwr.p.test(h=h, n=60)

    Spower(p_prop.test, n=60, prop=c(.5, .4))

}

}
