% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_r}
\alias{p_r}
\title{Correlation simulation and p-value}
\usage{
p_r(n, r, rho = 0, method = "pearson", two.tailed = TRUE)
}
\arguments{
\item{n}{sample size}

\item{r}{correlation}

\item{rho}{population coefficient to test against using
\code{\link[car]{linearHypothesis}}. Defaults to 0}

\item{method}{method to use to compute the correlation
(see \code{\link{cor.test}}). Only used when \code{rho = 0}}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}
}
\value{
a single p-value
}
\description{
Generates correlated X-Y data and returns a p-value to assess the null
of no correlation in the population. The X-Y data are generated
assuming a multivariate normal distribution.
}
\examples{

# 50 observations, .5 correlation
p_r(50, r=.5)
p_r(50, r=.5, method = 'spearman')

# test against constant other than rho = .6
p_r(50, .5, rho=.60)


if(FALSE){
    # compare simulated results to pwr package

    pwr::pwr.r.test(r=0.3, n=50)
    Spower(p_r, n=50, r=0.3)

    pwr::pwr.r.test(r=0.3, power=0.80)
    Spower(p_r, n=NA, r=0.3, power=.80, interval=c(10, 200))

    pwr::pwr.r.test(r=0.1, power=0.80)
    Spower(p_r, n=NA, r=0.1, power=.80, interval=c(200, 1000))

}

}
