% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_t.test}
\alias{p_t.test}
\title{Independent/paired samples t-test simulation and p-value}
\usage{
p_t.test(
  n,
  d,
  mu = 0,
  type = c("two.sample", "one.sample", "paired"),
  n2_n1 = 1,
  two.tailed = TRUE,
  raw_info = list(mu1 = NA, mu2 = NA, sigma1 = NA, sigma2 = NA)
)
}
\arguments{
\item{n}{sample size per group, assumed equal across groups}

\item{d}{Cohen's standardized effect size \code{d}}

\item{mu}{population mean to test against}

\item{type}{type of t-test to use; can be \code{'two.sample'},
\code{'one.sample'}, or \code{'paired'}}

\item{n2_n1}{allocation ratio reflecting the same size ratio.
Default of 1 sets the groups to be the same size. Only applicable
when \code{type = 'two.sample'}}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{raw_info}{(optional) list of mean and variance inputs for each group.
If supplied the \code{d} input will be ignored}
}
\value{
a single p-value
}
\description{
Function generates two sets of individuals with equal sample sizes
and Cohen's effect size 'd'. The data and associated t-test
assume that the conditional observations are normally distributed and have
have equal variance.
}
\examples{

# sample size of 50 per group, medium effect size
p_t.test(n=50, d=0.5)

# second group 2x as large as the first group
p_t.test(n=50, d=0.5, n2_n1 = 2)

# paired and one-sample tests
p_t.test(n=50, d=0.5, type = 'paired')
p_t.test(n=50, d=0.5, type = 'one.sample')

}
