% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_templates.R
\name{p_t.test}
\alias{p_t.test}
\title{Independent/paired samples t-test simulation and p-value}
\usage{
p_t.test(
  n,
  d,
  mu = 0,
  type = c("two.sample", "one.sample", "paired"),
  n2_n1 = 1,
  two.tailed = TRUE,
  var.equal = TRUE,
  raw_info = list(means = NA, sds = NA)
)
}
\arguments{
\item{n}{sample size per group, assumed equal across groups}

\item{d}{Cohen's standardized effect size \code{d}}

\item{mu}{population mean to test against}

\item{type}{type of t-test to use; can be \code{'two.sample'},
\code{'one.sample'}, or \code{'paired'}}

\item{n2_n1}{allocation ratio reflecting the same size ratio.
Default of 1 sets the groups to be the same size. Only applicable
when \code{type = 'two.sample'}}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{var.equal}{logical; use the classical or Welch corrected t-test?}

\item{raw_info}{(optional) list of mean and SD inputs for each group,
each specified as a vector. When specified the input \code{d} is ignored}
}
\value{
a single p-value
}
\description{
Generates one or two sets of continuous data group-level data
according to Cohen's effect size 'd', and return a p-value.
The data and associated t-test
assume that the conditional observations are normally distributed and have
have equal variance by default, however these may be modified.
}
\examples{

# sample size of 50 per group, "medium" effect size
p_t.test(n=50, d=0.5)

# second group 2x as large as the first group
p_t.test(n=50, d=0.5, n2_n1 = 2)

# paired and one-sample tests
p_t.test(n=50, d=0.5, type = 'paired')
p_t.test(n=50, d=0.5, type = 'one.sample')

}
