% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_sim_Gpower.R
\name{p_var.test}
\alias{p_var.test}
\title{p-value from variance test simulation}
\usage{
p_var.test(
  n,
  sds,
  n.ratios = rep(1, length(sds)),
  ratio = 1,
  sigma = 1,
  two.tailed = TRUE,
  test = "Levene",
  correct = TRUE
)
}
\arguments{
\item{n}{sample size per group, assumed equal across groups}

\item{sds}{a vector of standard deviations to use for each group}

\item{n.ratios}{allocation ratios reflecting the sample size ratios.
Default of 1 sets the groups to be the same size (n * n.ratio)}

\item{ratio}{hypothesized ratio of the population variance (only
used in two-sample case)}

\item{sigma}{standard deviation value to test against in one-sample test}

\item{two.tailed}{logical; should a two-tailed or one-tailed test be used?}

\item{test}{type of test to use in multi-sample applications.
Can be either \code{'Levene'} (default) or \code{'Bartlett'}}

\item{correct}{logical; use correction when \code{test = 'Bartlett'}?}
}
\value{
a single p-value
}
\description{
Generates one or or more sets of continuous data group-level data
to perform a variance test, and return a p-value. When two-samples
are investigated the \code{\link{var.test}} function will be used,
otherwise functions from the \code{EnvStats} package will be used.
}
\examples{

# one sample
p_var.test(100, sds=10, sigma=9)

# three sample
p_var.test(100, sds=c(10, 9, 11))

if(FALSE){
  # power to detect three-group variance differences
  Spower(p_var.test, n=100, sds=c(10,9,11))

  # sample size per group to achieve 80\% power
  Spower(p_var.test, n=NA, sds=c(10,9,11),
         power=.80, interval=c(100, 1000))
}

}
