% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spower.R
\name{Spower}
\alias{Spower}
\title{Simulation-based Power Analysis}
\usage{
Spower(
  conditions,
  sim_function,
  interval,
  power,
  sig.level = 0.05,
  replications = 10000,
  integer,
  beta_alpha = NULL,
  ...
)
}
\arguments{
\item{conditions}{a set of conditions to use in the simulation, passed to
either \code{\link{SimSolve}} if exactly one value contains an \code{NA} or to
\code{\link{runSimulation}} if \code{power} is set to \code{NA}}

\item{sim_function}{function that both creates the data and returns a single
p-value for the analysis of interest. Function must contain the arguments
\code{condition} to utilize the row in \code{conditions}}

\item{interval}{search interval to use when \code{\link{SimSolve}} is required}

\item{power}{power level to use. If set to \code{NA} then the empirical power
will be estimated given the fixed \code{conditions} input
(e.g., for post-hoc power analysis)}

\item{sig.level}{alpha level to use. If set to \code{NA} then the empirical
alpha will be estimated given the fixed \code{conditions} input
(e.g., for criterion power analysis)}

\item{replications}{number of replications to use when
\code{\link{runSimulation}} is required}

\item{integer}{a logical value indicating whether the search iterations
use integers or doubles.
Automatically set to \code{FALSE} if \code{interval} contains
non-integer numbers, though in general this should be set explicitly}

\item{beta_alpha}{ratio to use in compromise analyses corresponding to
the Type II errors (beta) over the Type I error (alpha). Ratios greater
than 1 indicate that Type I errors are worse than Type II, while ratios
less than one the opposite. A ratio equal to 1 gives an equal trade-off
between Type I and Type II errors}

\item{...}{additional parameters to pass to \code{\link{runSimulation}} or
\code{\link{SimSolve}}}

\item{Analyse}{analyse function (see \code{\link{runSimulation}})}
}
\description{
Description
}
